# frontend/openapi.json
{
...
  "info" : {
    "contact" : {
      "name" : "soptim",
      "url" : "https://www.soptim.de/"
    },
    "description" : "This API provides utilities for editing RDFGraphs that model UML classes using the CIM standard.",
    "license" : {
      "name" : "Apache License 2.0",
      "url" : "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "title" : "RDFArchitect backend",
    "version" : "0.0.0"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/datasets/{datasetName}/graphs/{graphURI}/shacl/custom/{shaclShapeURI}" : {
      "delete" : {
        "description" : "Delete a shacl shape form a shacl graph",
        "operationId" : "deleteShape",
        "parameters" : [
          {
            "description" : "The name/url of the inquirer.",
            "in" : "header",
            "name" : "Origin",
            "required" : false,
            "schema" : {
              "default" : "unknown",
              "type" : "string"
            }
          },
          {
            "description" : "The literal name of the dataset.",
            "in" : "path",
            "name" : "datasetName",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          },
          {
            "description" : "The url encoded uri of the graph, or \"default\" to access the default graph.",
            "in" : "path",
            "name" : "graphURI",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          },
          {
            "description" : "The property shape to delete.",
            "in" : "path",
            "name" : "shaclShapeURI",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          }
        ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "string"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "delete a shacl shape",
        "tags" : [
          "shacl"
        ]
      },
      "put" : {
        "description" : "Replace or insert a shacl shape from a shacl graph with the given triples in a valid turtle syntax.",
        "operationId" : "replaceShape",
        "parameters" : [
          {
            "description" : "The name/url of the inquirer.",
            "in" : "header",
            "name" : "Origin",
            "required" : false,
            "schema" : {
              "default" : "unknown",
              "type" : "string"
            }
          },
          {
            "description" : "The literal name of the dataset.",
            "in" : "path",
            "name" : "datasetName",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          },
          {
            "description" : "The url encoded uri of the graph, or \"default\" to access the default graph.",
            "in" : "path",
            "name" : "graphURI",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          },
          {
            "description" : "The property shape to replace.",
            "in" : "path",
            "name" : "shaclShapeURI",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          }
        ],
        "requestBody" : {
          "content" : {
            "text/plain" : {
              "schema" : {
                "description" : "the triples to replace, requires valid turtle syntax.",
                "type" : "string"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "string"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Replace/Insert shacl shape",
        "tags" : [
          "shacl"
        ]
      }
    },
...
  }
}